Tests de los features

+++ RealStateApp.Core.Application/Features/PropertyType/Commands/DeleteCommand/DeletePropertyTypeCommand.cs
@@ -1,4 +1,6 @@
+using System.Diagnostics.CodeAnalysis;
 using MediatR;
+using RealStateApp.Core.Application.Exceptions;
 using RealStateApp.Core.Domain.Interfaces;
 
 namespace RealStateApp.Core.Application.Features.PropertyType.Commands.DeleteCommand;
@@ -19,6 +21,9 @@ public class DeletePropertyTypeCommandHandler : IRequestHandler<DeletePropertyTy
 
     public async Task<Unit> Handle(DeletePropertyTypeCommand request, CancellationToken cancellationToken)
     {
+        var propertyType = await _propertyTypeRepository.GetByIdAsync(request.Id);
+        if (propertyType == null) throw new ApiException("property type not found");
+        
         await _propertyTypeRepository.DeleteAsync(request.Id);
         return Unit.Value;
     }
